rfi_pipeline: '_copy' version is ready, replace original
